Don't plot our position twice on the minimap.
Without this patch we get plotted along with all of the other avatars and then we get plotted again on top of our first marker.  This just saves the viewer a small amount of work and avoids the ghost effect where our second marker has to catch up with the first when we walk.