Added a RenderUseStreamVBO control checkbox on hardware preferences.  Also disabled RenderUseStreamVBO by default for fresh installations.
More often than not (it would seem), disabling this option results in a performance boost.  I have therefore disabled this option by default and added a checkbox for it in the the hardware preferences for those who find that it's of some benefit.

I also added even spacing between controls in the hardware preferences.  Some of the controls were bunched up while others were spread apart, and it all looked a bit messy - to me at least.