Restored RenderUseStreamVBO back to being enabled by default.
Thinking about my last patch, the tests I had performed were subjective and well within a margin of error so, on balance, it's probably best to leave the setting default as it was.  The user can run tests and determine whether to leave it enabled or to disable it.

Intel GPU users will probably want to disable this, whereas ATI and Nvidia users will see only a marginal difference either way.  The default for this setting should probably be specified in the GPU table.