Added the ability to pan the world map using the middle mouse button.
Users can still use the left mouse button to do the same thing as usual if they prefer.  I'm just used to zooming and panning with the scroll wheel and selecting with the left mouse button so this change is for me really.  It ties it all in with the similar I change made to the mini-map as well.