Added a new "Permissions" sub-menu for friends on the People floater.
The usual per-friend permissions are provided, as follows:

    * See when I am online
    * Find me on the world map
    * Edit, delete or take my objects

It takes a while for the server to process the request so if you make multiple changes for the same avatar then later changes may override the earlier ones.  I added a notification toast to slow people down a bit.

Note: The checkboxes on the legacy profile look quicker because it fakes the tick in advance of the server processing the request.  We can't do that in the menu because the actual status will override any fake settings as soon as we re-open the sub-menu.  This checkbox faking is also the reason why the checkboxes don't update in real time when you have the profile open at the same time as you set permissions in the People panel.  Close the legacy profile floater and re-open, and it will re-read the actual permission status.