Added a button to copy the size, rotation and offset material settings from the diffuse map to both the normal and specuar maps (if the extra materials are present).
Also added a button to copy from normal to both diffuse and specular, and another to copy from specular to both diffuse and normal.

Enter the details once into whichever material pane you are looking at, and then hit the "Duplicate Values" button to copy the values to the other two material types.