Defer the request for the avatar's age in the name tag until we have the avatar's name (if the age is to be displayed at all, that is).
This avoids a request collision leading to a tag with nothing but an age in it.  This was happening mostly with "joined today" players who were unknown to the sim's name caches.

I found this bug and verified the fix while watching a landing point in an area with lots of brand new arrivals (Hyde Park in London).