Added a "Web Browser" toybox button and associated icon.
Added a new WebBrowserHomePage setting (defauts to http://www.google.com/).

Added the new WebBrowserHomePage setting to the Preferences floater.

Modified the web Content floater to add a "Home" button that sends the user to their preferred home page.

TODO: Add a button to save the current browsed-to page as the home page.