Allow land owners/managers to specify the region restart delay.
The restart delay was hard-coded to 120 seconds previously.  Admins may now specify any grace period between 15 seconds and one hour.  The default for fresh viewer installations is five minutes.

I decided against allowing the minimum grace period to be zero seconds because it just seemed silly to do so.  15 seconds is therefore the minimum.

The upper limit is probably far too large, but I thought that any larger value would be pointless.  An hour seemed to be a reasonable, although still somewhat excessive, upper limit.  It just had the advantage of being a nice round number.

If someone really wants to use an out of bounds number then they can monkey about with the debug settings (RestartDelaySeconds variable) to achieve this.  Don't blame me if you do this and it all ends in tears. :-)

Feature requested by Jason Keegan.