Added a "Share" option to the avatar context menu.
Yes, you can drop items on the avatar, but they may be in a difficult to target dance animation.  Yes, you can drop items on their name tag or start an IM and drop items in there.  This just gives another option, and makes the function clear in the menu from a usabiliy standpoint.