Added the ability to share landmarks from the Places floater.
The only way to share landmarks previously was to find them in inventory and hand them out from there.  I don't know of any other viewer that allows landmarks to be shared directly from the Places floater.

As per the usual sharing rules, either one or more landmarks may be shared, or exactly one folder full of items (landmarks in this case) may be shared.  We cannot select two or more folders, or a mix of folders and individual landmarks.  If there's an upper limit on the number of people who can be selected to receive items in one share operation then I don't know what that limit is, and haven't taken it into account.  Obviously, the lower limit is one.

There are no permission checks to perform when with sharing landmarks;  Landmarks always have full permissions.  Please correct me if I'm wrong. :-)