First cut at a context menu for Area Search.
All works fine apart from the "Take" and "Buy" options.  The enabler code is broken somehow and I haven't been able to spot what's wrong with it yet.  I've commented out the XML for those two menu items for now, but they will go back once the code is fixed.  The take/buy operations work fine, by the way - it's just the menu item enabler.  I can't see anyone wanting to do a long distance "blind buy" anyway, but the "take" option needs to be enabled when appropriate.

I've left a couple of minor TODO comments in the code, and I'll revisit those.

The next step will be to redesign the floater itself to add options and to enhance its usability.