Stream reliability patch (hopefully).
This patch is in response to reports of losing the stream and being unable to restart it.  My thinking is there is a difference between the actual play state and what the viewer thinks is going on.  This patch *should* force the stream to restart whenever it's asked to play.

If this works then it's just masking an underlying problem somewhere.  On the other hand, if this works then finding the underlying problem won't be so much of a priority.

I'm not really the best person to be looking into stream-related problems. :-)