Set a limit on excessive geometry to avoid graphics lag.
This applies generally to park areas with excessive amounts of high-geometry sculpted grass, where the gardener has layered several of these objects over one another in order to heavily blanket the area and almost completely hide the underlying terrain texture.  Some of the excess geometry will be automatically derendered in order to keep the frame rate from crashing down to a tiny fraction of its usual value.

This particular performance killer has taken a while to track down, and this appears to be the best solution to guard against clusters of objects with excessive geometry dragging the rendered FPS down to a crawl.  Only the griefer-class clustered geometry bombs will be targeted;  The rest of the scene will render normally, and a sane amount of grass blanketing the area will render perfectly as well.  There's just too much work for the shaders to do when people go over the top.