Made the geometry overload limit (changeset 41bbeda) optional.
A new control on the Preferences floater's Graphics tab allows the user to switch geometry overload protection on or off.  The default is to have the overload protection switched on.

It was suggested to me that the protection should be optional because offending objects cannot be deleted, ARed or returned if they cannot be seen.  Users without such powers will most likely want to leave the protection switched on.

It was also suggested that the protection be extended to cover avatars who TP in wearing griefer-sized geometry in an attempt to knock everyone's FPS down to a crawl.

I tweaked the proximal object group setting a little (up from 120 to 130) after some experimentation.  The avatar limit may need some tweaking, but the initial value chosen (2500) seems high enough to not come anywhere near affecting normal users.  Avatars exceeding the limit will be rendered as imposters.