Area search rewrite with a shiny new UI design for enhanced usability.
Also replaced the toolbar icon image.  I wasn't really happy with the last one I made.  This one was drawn to look much like the existing binoculars icon used for the draw distance setting.

The options menu at the lower left will hold more facilties in the future, such as toggling regex search mode.  There's only really one way to do regex searches so the code for that might as well be copied from Firestorm.

Note: The licence has been changed to the LGPL v2.1 going forward.  Yes - the previous BSD-style licence allows that as long as the terms of the original licence are also complied with.  This change ensures that future versions of this code will continue to be free software.

Still to do:

    * Regular expression searches

    * Support for multiple selections

    * Still haven't looked into why the XXX and YYY options don't get enabled (they are still commented out in the XML until they get looked into and fixed)