Fixed the new texture preview/picker Alpha/Opaque display.
I used the wrong technique to determine whether the texture had an alpha channel.  It worked in all of my tests and then, upon release, I was promptly shown a texture that made it fail.  Lovely.

Also hide the "[WIDTH]x[HEIGHT] ([ALPHA])" text on the texture preview floater until we have information to display.  This avoids raw tokens being displayed for a short time if the image takes a while to get to us.