Moved the name tag preference lookups out of llviewerdisplay and into llvoavatar.
It was wasteful to do the lookups with every frame, as the settings rarely change.  It's better to use a local static LLCachedControl in the LLVOAvatar class, right where they are used.

Committed separately to the previous changeset (f3d30c2ef77e) that converted gSavedSettings.getFoo() to static LLCachedControls, so that this may be reverted if it causes problems.  My tests indicate that this change is for the better.