Final part of the status bar layout overhaul.
The previous four changesets have been leading up to this.  We can now properly hide the bandwidth graph and the new FPS counter, rather than just rendering them invisible.  I've also added a setting to allow the draw distance slider to be hidden.

Right click on the status bar to see the new options in the menu.

The new FPS display is disabled by default for fresh installations because it's a bit distracting to see it whizzing away up there.  If people want to see it then they can turn it on.  I haven't added any of this to the Preferences floater, as I think the status bar context menu is a better place for the settings.  Let me know if you disagree.