Fixed what looks like a merge error in the avatar context menu in chat.
This prevented "Zoom In" and a couple of other menu items from working.

It seems that "Block Text" and "Block Voice" are two separate choices now, so that's been fixed up as well.  This change may have been the source of the merge error.

Also added the new "Request Teleport" facility to the avatar chat context menu, as well as a "Show On Map" option (friend permissions notwithstanding).

TODO: I just spotted that "map" and "teleport" in LLChatHistoryHeader::onObjectIconContextMenuItemClicked() have hard-coded "secondlife://" SLURLs that may need to be changed to work with non-SL grids.  That's separate from this commit;  This one deals with the Avatar context menu.