Fixed a bug in the bandwidth graph widget display.
A raw bandwidth usage snapshot was being used in the graph, rather than the "per second" rate.

This was either introduced when I changed the widget to be definable in XML, or it existed prior to that and I just never noticed before.  I'll have to review my previous changeset to find out, but never mind where it came from - it's fixed now and the graph does something useful for a change.

Also, the maximum bandwidth was using "getMaxBandwidth() / 1000" where "... * 1024" is used to set the value.  1000 is actually correct in modern metric Si units, but as the rest of the viewer incorrectly uses 1024 for kilo so the code was inconsistent there.  It's probably been like that for ages, but as the graph didn't actually work, nobody noticed.  Your broadband Internet feed is metered in Si units, by the way so a kilobit is 1000 bits and a megabit is 1000 kilobits etc.  Your hard disk size is measured that way too.  It seems only memory is measured in non-Si units these days, for technical reasons.