Don't show SL-specific menu options when not in SL.
Added a new only_in_sl="true" parameter to the XUI, which can be used with all XUI elements.

If you are not in SL and the XUI element has the only_in_sl parameter set true, then the element will be disabled and invisible regardless of any other visibility settings.

The following main menu items are now disabled and invisible when not in SL:

Me:

    * Buy L$
    * Merchant Outbox
    * Account Dashboard

Build:

    * Linksets
    * Pathfinding

Help:

    * Wiki
    * Community Forums
    * Support Portal


The following are also disabled and invisible when not in SL:

    * The object context menu's "Show in linksets" and "Show in characters" options (pathfinding is SL-specific).
    * The Inventory floater's "Shop" button (the marketplace website is SL-specific).

Note that any on_enable/on_visible callbacks that have been defined will not be called if the element is disabled via the only_in_sl option.


If I missed anything then tack a comment onto this changeset.


Yet more items for the todo list:

    * Replace the Inventory floater's "Shop" button altogether so it doesn't leave a gap on non-SL grids.

    * Sometime in the future, change only_in_sl="true" to something more like if_grid="sl|opensim|aurora", where one or more sim types can be specified to check against.  Maybe even supporting if_grid="!sl" for "if not in SL".  There's no need for this right now, so the simplest method is best.