Added a "Camera mouse sensitivity" setting to the move/view preferenecs panel.
Somewhere between 3.7.1 and 3.7.4 the mouse-controlled camera got very slow.  I suspect there's a mixup between degrees and radians somewhere in the code, but I can't find it.  I skipped a few versions, so I don't know which specific verson changed the mouse sensitivity.

This new setting acts as a multiplier, and will still be useful as a preference even if the underlying issue is fixed upstream.  The preference range may need to be modified if upstream make the mouse-controlled camera more sensitive again though.