Name tags were sometimes showing when they shouldn't and sometimes wouldn't fade in "show briefly" mode.
Problems spotted when testing a another patch.  I abandoned the patch I was testing, but the problems spotted needed to be corrected.

LLVOAvatar::sRenderName was being set to the NameTagShowGroupTitles preference in some cases so the AvatarNameTagMode preference was being overridden in some instances.  Tould be made to happen by toggling name tags off/on while also toggling group tags off/on in various combinations.

LLVOAvatar::sRenderName actually had no value and was also preventing name tag fadeout in "show briefly" mode in some instances.  Conveted the few uses to use a cached lookup of the AvatarNameTagMode preference instead, and deleted the variable.

Also, LLVOAvatar::sRenderGroupTitles was unused apart from in some commented out code.  Deleted the variable and the commented out code that was littering the place and causing code readability-related confusion.