The Inventory Filter floater now has a title.
It looked to me as if the title was supposed to be read from the filter text input area on the main Inventory floater, but this wasn't happening.  I didn't see any value in that, so rather than trace and fix it, I just set a static "FILTER INVENTORY" title in the XML and removed the title override code.