Can now specify how many per-pixel samples are taken when anisotropic filtering is enabled.
Previously the viewer always used the maximum value supported by the GPU when enabled, so it was an "all or nothing" setting with a simple checkbox in preferences.

Note: Users who had anisotropic filtering enabled using a viewer released before this changeset was applied will find that it is now disabled, and will need to be re-enabled via the Hardware floater in Preferences.  The new selection will persist going forward.