We now get an "Inventory Accepted/Denied" notification toast when we share items, like we used to.
It seemed to me that log_to_im="true" was opening an IM session to write the share notification message and then show_toast="true" was aborting because there was an IM session open.

I swapped the blocks of code responsible for the above so it first checks for an open IM session before displaying the notification.  Once that's done it logs to IM (opening a new one if necessary).

Side note: It seems to me as if there's a bunch of RLVa code in the notification system that's stomping around without checking whether RLVa is enabled or not.  I'll have to look into that at some point.  The RLVa code is so difficult to read, follow and maintain (especially with all the graffiti that comes with it), so I'll leave that for when I have a year or so to decypher it all.