Added the grant/revoke friend permissions sub-menu to the avatar context menu to match the one I added to the the friends list a long time ago.
If a friend is in sight then we should be able to click on the avatar to get those options instead of having to look them up in the friends list.

The code I added to handle friend permissions has been moved to the LLAvatarActions class to allow its re-use in multiple menus.  This sub-menu will be added to the chat context menu in one of the next commits.

Also, context sub-menus will now be hidden from the parent context menu if there are no visible items in them.  This was needed in order to hide the "Permissions" sub-menu when right-clicking on avatars who are not in your friends list.