Made the chat text entry widget's background colour differ according to the chat type.
Nearby chat remains as white.  The rest of the chat types are coloured as follows:

    * IM sessions are tinted red.
    * Group chat sessions are tinted green.
    * Conference sessions are tinted blue.

Several users have asked me to find a way to help prevent them from typing into
IMs and group chats by mistake when they meant to be typing into nearby chat
etc.  This is an experiment to see if the extra visual cue helps.

This can be switched on and off using a new checkbox in the "Chat" tab in the
Viewer preferences panel.
This new setting is switched on by default for the next release but, depending
upon user feedback, I may change my mind about the default for future releases.