Fixed the avatar cloud issue.
It was just a merge error in the end, but it was a nightmare to pinpoint the code that was failing.  Once the railure point had been determined, it was easy to diff the area to see what was missing.  This merge error was similar to the last one, so I'm going to check the recent merges for similar errors.  I checked already after I spotted the last one, but clearly missed this one.  I'll check again.

As suspected, the avatar cloud and the duplicate entries in the worn items and outfits lists are seperate issues.  The cloud is fixed now, but the duplicates remain.

The duplicate entries are most likely just a display issue and won't affect further testing of the work to date, so ignore the duplicates, test the menus and other things and I'll track that down in parallel with work others are doing with this merge.

I've verified that I decloud.  I took off all my attachments to test with, so I've verified that wearing and adding works.  Further tests are needed, and I haven't tried wearing and adding clothing items or changing into a new skin/shape etc.  This all needs to be tested.  Pretend the duplicates are not in the list for now, as they should just be ghosts.  If you see a triplicate then that could be a problem.