DRTVWR-418: Use separate XML string elements for command arguments.
autobuild now passes a vector of arguments to the commands it executes,
instead of smashing everything together into a single string only to have
cmd.exe parse it back into individual arguments again. In general, this is a
good thing and more robust than before.

However, it surfaces the fact that sometimes people were lazy and encoded
things like "-configuration Release" or "-j 12" as single <string> elements in
the XML. When such things are handed to the command as a single argument string
containing a space, it can get confused.

Code (e.g.) <string>-configuration</string><string>Release</string> instead.