Ticket# 490 SL - x/y/z positions still snap to a maximum in edit window when entered or nudged
Backout of (b7a845597013) Was intended to be backed out earlier, but was not.
Making a number of adjustments to better position objects on large OpenSim
regions, including the ability to negative Z value (clamped at -256 for now).
The viewer will now respect the SecondLife minimum Z value of 0 (zero).
A better mechanism is needed to retrieve the exact dimensions of an OpenSim
region to set the correct values.  More testing on var regions are needed.