Marine's commit 31569ccc229d presupposed the CanOOC functionality is no longer required in line with her current direction with the hardcore viewer that doesn't allow CanOOC to be turned on (put another way, that commit breaks it by no longer routing (( )) to normal chat where it might get stamped on by crunchemote if CanOOC was not enabled).
This commit retains the change Marine intended with (( )) chat now getting
sent through the redirection despatcher but also allows it to fall through
and appear in chat if CanOOC is allowed (and the (( )) syntax is matched).

The only downside is that OOC speech may get seen twice (once through redirected
chat and once in normal chat). Since we check for sCanOoc we'll only fall through
to normal chat when OOC is permitted and won't have the old behaviour of ...
appearing in chat for disallowed OOC speech.