DAYTW-34 Replace all occurrences of kokua_icon.icns and kokua_icon.bmp with the corresponding Dayturn icon versions.
The BMP versions are not used by the macOS version of the viewer and shall be removed. There are also remains fo SecondLife icon files to be removed.
TODO: Add the macOS asset catalog of icons to the project and viewer for full support of both Retina displays, and the small and large icons in the Finder.

# Conflicts:
#	indra/newview/CMakeLists.txt
#	indra/newview/icons/release/kokua_icon.bmp
#	indra/newview/icons/release/kokua_icon.icns
#	indra/newview/icons/release/kokua_icon.ico
#	indra/newview/icons/release/kokua_icon.png
#	indra/newview/icons/test/kokua_icon.bmp
#	indra/newview/icons/test/kokua_icon.icns
#	indra/newview/icons/test/kokua_icon.ico
#	indra/newview/icons/test/kokua_icon.png
#	indra/newview/kokua_icon.icns
#	indra/newview/viewer_manifest.py

# Conflicts:
#	indra/newview/viewer_manifest.py